requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Machine Learning
xgboost>=1.7.0
lightgbm>=4.0.0
catboost>=1.2.0

# Time series and advanced ML
tsfresh>=0.20.0
umap-learn>=0.5.0

# Visualization and demos
streamlit>=1.25.0
gradio>=3.40.0

# Configuration and utilities
pyyaml>=6.0
omegaconf>=2.3.0
tqdm>=4.65.0
joblib>=1.3.0

# Explainability
shap>=0.42.0
lime>=0.2.0

# Development and testing
pytest>=7.4.0
black>=23.0.0
ruff>=0.0.280
mypy>=1.5.0

# Optional: Advanced features
# torch>=2.0.0
# transformers>=4.30.0
# mlflow>=2.5.0
